+++ views/admin/info/index.info.php
@@ -10,6 +10,7 @@
             <ul class="uk-nav uk-nav-side"
                 data-uk-switcher="{connect:'#search-info-content',swiping:false, active:0}">
                 <li><a href="#">Overview</a></li>
+                <li><a href="#">Important</a></li>
                 <li><a href="#">Search widget</a></li>
                 <li><a href="#">Driven Listings Plugin</a></li>
                 <!-- <li><a href="#">Data Statistic</a></li> -->
@@ -28,6 +29,10 @@
 
                 <li>
                     <?php include 'overview.php'; ?>
+                </li>
+
+                <li>
+                    <?php include 'important.php'; ?>
                 </li>
 
                 <li>